Re-organized directory structure and added testing framework.
Because this will be written in several languages, the package structure has been chaned to have a "lang" folder at the root. The only language at this time is JavaScript.

A testing framework has been added, RhinoUnit, and several tests have been added. Note: The tests are not hooked into the framework yet, so they will not work. Note: The LINT checker is throwing an exception. I am not sure what is causing it, but I have run the code against test scenarios locally and have not yet found a problem. Hopefully, hooking the tests into the testing framework will shine a greater light on what is wrong.