Added tests and builds.
The test definitions and data have been completed. The testing framework and tests have been built for JavaScript.

There is now a root build file as well as a build file for JavaScript that has two targets, "dist" and "clean". JavaScript has an additional target, "dist-min", which will minify the output JavaScript.