Renamed the schema field names.
When defining a sub-schema for an object or an array, the field name "schema" was used. This breaks our own rule of not having a single field with multiple types, and it was mildly confusing to have "schema" mean different things depending on its type and the value of its correspoding "type" field.

Now, when defining an object, the accompanying key is "fields", which is still an array of objects where each object defines a single field. When defining a constant-type array, the accompanying key is "constType", which is still an object defining the type for all indicies in the corresponding data array. When defining a constant-length array, the accompanying key is "constLength", which is still an rray of objects where each object defines the type for that index in the corresponding data array.

There is a chance that constant-length arrays will be removed altogether. In that case, along with removing this type, constant-length fields will just be referred to as arrays and their accompanying key will, again, be renamed to something like "indicies" or "indexType".