Added the Java implementation.
The Java implementation is complete. The majority of the tests are there including all of the core ones that test schema and data validation. There are some extra tests that exist in the JavaScript module that have not yet been imported.

Some tests were modified to better test the system.

A minor fix was made to the JavaScript implementation to reflect the updated test.

This should complete release 0.3.