Fixed a bug regarding how missing fields were serialized.
To prevent null fields from being serialized, an annotation, @JsonInclude(), was being used. The problem is that the inclusion flag being used was, Include.NON_DEFAULT, which required a private default constructor to initialize the fields to their "default" values. This was unnecessary because the only time it shouldn't have been included was when it was null, so the inclusion flag Include.NON_NULL is now being used.